java/Makefile.am: fix build with Java 8
The patch with added support for Java 11 accidentally broke
java 8 support, when it added $(abs_srcdir) to the *.java
files at zbar_jar_SRCS. Revert this change and rework on
Java 11 logic.

The end result is that now the Java 11 logic is cleaner and
easier to maintain. It will also cleanup the "javac -h ."
mess of creating several individual headers that we don't
want at the builddir.

While here, be sure that the produced zbarjni.h file will
be cleaned at "make clean".

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>